pyproject.toml
[project]
name = "orange-hrm-playwright"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "playwright>=1.58.0",
    "pylint>=4.0.5",
    "pytest-playwright>=0.7.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["pages*"]

[tool.pylint.messages_control]
disable = ["missing-module-docstring", "missing-function-docstring", "too-few-public-methods"]
